+++ src/service/scheduleService.js
@@ -1,12 +1,11 @@
 import cloneDeep from 'lodash.clonedeep';
 import sha256 from 'crypto-js/sha256';
 import Md5 from 'crypto-js/md5';
-import { RRule, datetime } from 'rrule';
 import Base64 from 'crypto-js/enc-base64';
 import isPublished from '../util/isPublished';
 import Logger from '../logger/logger';
-import ConfigLoader from '../config-loader';
-import ScheduleUtils from "../schedule";
+import ConfigLoader from '../util/config-loader';
+import ScheduleUtils from '../util/schedule';
 
 /**
  * ScheduleService.
@@ -190,7 +189,10 @@ class ScheduleService {
 
         // Run through all schedule item and see if it occurs now. If one or more occur now, the playlist is active.
         schedules.every((schedule) => {
-          const scheduleOccurs = ScheduleUtils.occursNow(schedule.rrule, schedule.duration);
+          const scheduleOccurs = ScheduleUtils.occursNow(
+            schedule.rrule,
+            schedule.duration
+          );
 
           if (scheduleOccurs) {
             active = true;